requirements.txt
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
duckdb>=0.9.0
pyarrow>=13.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
umap-learn>=0.5.4
wandb>=0.16.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
tqdm>=4.66.0
pyyaml>=6.0
